server/fileServer/requirements.txt
grpcio==1.46.0
protobuf==3.20.1
grpcio-tools==1.46.0
yt-dlp==2023.2.17